pyproject.toml
[project]
name = "pytest-dbt-duckdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "absl-py>=2.1.0",
    "dbt-duckdb>=1.9.1",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "pytest>=8.3.4",
    "pytest-env>=1.1.5",
    "ruamel-yaml>=0.18.10",
]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pre-commit>=4.1.0",
    "ruff>=0.9.3",
    "tox>=4.24.1",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.50",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[tool.mypy]
disable_error_code = ["call-arg", "import-untyped"]
